add beerDetail page

+++ src/main/java/com/techelevator/clebrews/model/BreweryDAO.java
@@ -10,18 +10,18 @@ public interface BreweryDAO {
 	public void saveBrewery(String name, String address, String city, String zipcode, String phoneNumber, String description, String breweryLogoUrl, String imgUrl, String websiteUrl, String businessHour);
 
 	public void updateBrewery(String name, String address, String city, String zipcode, String phoneNumber,
-			String description, String breweryLogoUrl, String imgUrl, String websiteUrl, String businessHours, int id);
+			String description, String breweryLogoUrl, String imgUrl, String websiteUrl, String businessHours, long id);
 	
 	public boolean searchForBrewery(String name);
 	
-	public Brewery getBreweryById(int id);
+	public Brewery getBreweryById(long id);
 
 
 	public Brewery getBreweryByUserId(int userId);
 	
 	public List<Brewery> getBreweryWhereUserIdIsNULL();
 
-	public void updateBreweryUserId(int id, int userId);
+	public void updateBreweryUserId(long id, int userId);
